+++ CHANGELOG.md
@@ -9,6 +9,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 * Improved hierachy of H1, H2, H3 etc. (accessibility)
 
+## v1.41.1 - 14th November 2018
+
+### Modified
+
+* Changed items to be lists
+* changed postcode to be paragraph instead of Heading
+
 ## v1.41.0 - 14th November 2018
 
 ### Added